@@ -6,90 +6,88 @@
   "scripts": {
     "start": "react-app-rewired start",
     "build": "react-app-rewired build",
-    "test:types": "tsc",
-    "test:unit": "react-app-rewired test --modulePaths=src",
+    "test": "react-app-rewired test --modulePaths=src",
     "test:lint": "eslint src --ext .js,.jsx,.ts,.tsx",
-    "presetup": "npm install cli-spinner prompt colors --no-optional",
+    "presetup": "npm install cli-spinner prompt colors --no-optional && cp .env.sample .env",
     "setup": "node config/setup/setup.js",
     "release": "release-it -c config/.release-it.json"
   },
   "dependencies": {
     "@adaptagency/fe_tools": "^1.6.4",
-    "@types/jest": "^24.0.15",
-    "@types/node": "^12.0.8",
-    "@types/react": "^16.8.22",
-    "@types/react-dom": "^16.8.4",
+    "@babel/preset-env": "^7.6.2",
+    "@loadable/component": "^5.10.3",
+    "@types/http-proxy-middleware": "^0.19.3",
+    "@types/jest": "^24.0.18",
+    "@types/node": "^12.7.8",
+    "@types/react": "^16.9.3",
+    "@types/react-dom": "^16.9.1",
     "@types/react-loadable": "^5.5.1",
-    "@types/react-redux": "^7.1.0",
-    "circular-dependency-plugin": "^5.0.2",
-    "critters-webpack-plugin": "^2.3.0",
+    "@types/react-redux": "^7.1.4",
+    "autoprefixer": "^9.6.1",
+    "circular-dependency-plugin": "^5.2.0",
+    "clean-webpack-plugin": "^3.0.0",
+    "compression-webpack-plugin": "^3.0.0",
+    "critters-webpack-plugin": "^2.4.0",
+    "css-loader": "^3.2.0",
     "cuid": "^2.1.6",
-    "eslint-config-prettier": "^5.1.0",
+    "customize-cra": "^0.8.0",
+    "eslint-config-prettier": "^6.3.0",
     "express": "^4.17.1",
-    "extract-css-chunks-webpack-plugin": "^4.5.2",
-    "node-sass": "^4.12.0",
-    "prerender-loader": "^1.3.0",
-    "react": "16.8.6",
-    "react-dom": "16.8.6",
-    "react-loadable": "^5.5.0",
-    "react-redux": "^5.0.7",
-    "react-router-config": "^4.4.0-beta.1",
-    "react-router-dom": "^4.3.1",
-    "react-scripts": "^3.0.1",
-    "redux": "^4.0.1",
-    "redux-actions": "^2.6.3",
-    "redux-observable": "^1.1.0",
-    "redux-persist": "^5.10.0",
-    "redux-promise": "^0.6.0",
-    "redux-thunk": "^2.3.0",
-    "style-loader": "^0.23.1",
-    "typesafe-actions": "^4.4.2",
-    "typescript": "3.4.3"
-  },
-  "devDependencies": {
-    "@hot-loader/react-dom": "16.8.6",
-    "@svgr/webpack": "^4.3.0",
-    "@types/http-proxy-middleware": "^0.19.2",
-    "@typescript-eslint/eslint-plugin": "^1.11.0",
-    "@typescript-eslint/parser": "^1.11.0",
-    "autoprefixer": "^9.6.0",
-    "css-loader": "^3.0.0",
-    "customize-cra": "^0.2.13",
-    "eslint": "^5.16.0",
-    "eslint-config-airbnb": "^17.1.0",
-    "eslint-config-react-app": "^4.0.1",
-    "eslint-plugin-import": "^2.17.3",
-    "eslint-plugin-jsx-a11y": "^6.2.1",
-    "eslint-plugin-react": "^7.13.0",
-    "eslint-plugin-react-hooks": "^1.6.1",
+    "extract-css-chunks-webpack-plugin": "^4.6.0",
     "html-webpack-plugin": "^3.2.0",
-    "http-proxy-middleware": "^0.19.1",
-    "husky": "^2.5.0",
+    "http-proxy-middleware": "^0.20.0",
+    "node-sass": "^4.12.0",
     "postcss-calc": "^7.0.1",
     "postcss-flexbugs-fixes": "^4.1.0",
-    "postcss-inline-svg": "^4.0.0",
+    "postcss-inline-svg": "^4.1.0",
     "postcss-loader": "^3.0.0",
+    "prerender-loader": "^1.3.0",
+    "react": "16.10.1",
+    "react-app-rewired": "^2.1.3",
+    "react-dom": "16.10.1",
+    "react-redux": "^7.1.1",
+    "react-router-config": "^5.1.1",
+    "react-router-dom": "^5.1.1",
+    "react-scripts": "^3.1.2",
+    "redux": "^4.0.4",
+    "redux-actions": "^2.6.5",
+    "redux-observable": "^1.2.0",
+    "redux-persist": "^6.0.0",
+    "redux-promise": "^0.6.0",
+    "redux-thunk": "^2.3.0",
+    "sass-loader": "^8.0.0",
+    "sass-resources-loader": "^2.0.1",
+    "style-loader": "^1.0.0",
+    "typesafe-actions": "^4.4.2",
+    "typescript": "3.6.3",
+    "webpack": "^4.41.0"
+  },
+  "devDependencies": {
+    "@hot-loader/react-dom": "16.9.0",
+    "@svgr/webpack": "^4.3.3",
+    "@typescript-eslint/eslint-plugin": "^2.3.1",
+    "@typescript-eslint/parser": "^2.3.1",
+    "eslint": "^6.5.0",
+    "eslint-config-airbnb": "^18.0.1",
+    "eslint-import-resolver-typescript": "^1.1.1",
+    "eslint-plugin-import": "^2.18.2",
+    "eslint-plugin-json": "^1.4.0",
+    "eslint-plugin-jsx-a11y": "^6.2.3",
+    "eslint-plugin-prettier": "^3.1.1",
+    "eslint-plugin-react": "^7.14.3",
+    "husky": "^3.0.7",
+    "prettier": "^1.18.2",
     "react-app-rewire-eslint": "^0.2.3",
     "react-app-rewire-hot-loader": "^2.0.1",
-    "react-app-rewire-typescript": "^2.0.2",
-    "react-app-rewired": "^2.1.3",
-    "react-hot-loader": "^4.11.1",
-    "release-it": "^12.3.0",
-    "sass-loader": "^7.1.0",
-    "sass-resources-loader": "^2.0.1",
-    "uglifyjs-webpack-plugin": "^2.1.3",
-    "webpack": "^4.34.0",
-    "webpack-bundle-analyzer": "^3.3.2",
-    "webpack-delete-after-emit": "^1.0.5",
-    "webpackbar": "^3.2.0"
+    "react-hot-loader": "^4.12.14",
+    "release-it": "^12.4.1",
+    "webpack-bundle-analyzer": "^3.5.2",
+    "webpackbar": "^4.0.0"
   },
   "sideEffects": [
     "*.css",
     "*.scss"
   ],
-  "eslintConfig": {
-    "extends": "react-app"
-  },
   "browserslist": [
     "> 1%",
     "last 2 versions",
