@@ -1,4 +1,4 @@
-import * as logger from "./logger-service";
+import * as logger from './logger-service';
 
 export default {
   logger,
