@@ -1,3 +1,3 @@
-declare module "RootTypes" {
-  export type Services = typeof import("./index").default;
+declare module 'RootTypes' {
+  export type Services = typeof import('./index').default;
 }
