@@ -1,5 +1,5 @@
-import foo from "state/ducks/foo";
+import { actions as fooActions } from 'state/ducks/foo';
 
 export default {
-  foo: foo.actions,
+  foo: fooActions,
 };
