@@ -1,6 +1,6 @@
-import { combineReducers } from "redux";
+import { combineReducers } from 'redux';
 
-import foo from "state/ducks/foo";
+import foo from 'state/ducks/foo';
 
 const rootReducer = combineReducers({
   foo: foo.reducer,
