@@ -5,7 +5,9 @@ interface Window {
 }
 
 // CSS Modules Styling
-declare module "*.scss" {
+declare module '*.scss' {
   const styles: { [className: string]: string };
   export default styles;
 }
+
+declare const require: any;
