@@ -1,4 +1,4 @@
 // Read more about Typescript Redux setup here
 // https://github.com/piotrwitek/react-redux-typescript-guide#store-configuration
 
-declare module "RootTypes";
+declare module 'RootTypes';
