@@ -1,7 +1,4 @@
 import promise from 'redux-promise';
 import thunk from 'redux-thunk';
 
-export const middlewares = [
-  thunk,
-  promise,
-];
+export default [thunk, promise];
