@@ -1,4 +1,4 @@
-import React from "react";
+import React from 'react';
 
 const Footer = (): JSX.Element => <footer className="footer" />;
 
