@@ -1,20 +1,20 @@
-import React from "react";
-import Types from "RootTypes";
-// import { Dispatch, bindActionCreators } from "redux";
-import { connect } from "react-redux";
+import React from 'react';
+import Types from 'RootTypes';
+import { Dispatch } from 'redux';
+import { connect } from 'react-redux';
 
 // Actions
-import { actions as fooActions } from "state/ducks/foo";
+import { actions as fooActions } from 'state/ducks/foo';
 
 const mapStateToProps = (state: Types.RootState) => ({
   title: state.foo.title,
   elements: state.foo.elements,
 });
 
-const mapDispatchToProps = dispatch => ({
-  addElement: (value: string) => dispatch(fooActions.addElement(value)),
-  updateTitle: (value: string) => dispatch(fooActions.updateTitle(value)),
-  deleteElement: () => dispatch(fooActions.deleteElement()),
+const mapDispatchToProps = (dispatch: Function) => ({
+  addElement: (value: string): Dispatch => dispatch(fooActions.addElement(value)),
+  updateTitle: (value: string): Dispatch => dispatch(fooActions.updateTitle(value)),
+  deleteElement: (): Dispatch => dispatch(fooActions.deleteElement()),
 });
 
 // Collecting the props into one type you can refer to.
@@ -26,45 +26,43 @@ const Frontpage: React.FC<Props> = ({
   updateTitle,
   addElement,
   deleteElement,
-}) => {
-  return (
-    <div className="page page-frontpage">
-      <h1>Home {title}</h1>
+}) => (
+  <div className="page page-frontpage">
+    <h1>Home {title}</h1>
 
-      <label htmlFor="idChanger">
-        Change the title
-        <input
-          type="text"
-          id="idChanger"
-          onChange={(e: React.ChangeEvent<HTMLInputElement>) => updateTitle(e.target.value)}
-        />
-      </label>
+    <label htmlFor="idChanger">
+      Change the title
+      <input
+        type="text"
+        id="idChanger"
+        onChange={(e: React.ChangeEvent<HTMLInputElement>): Dispatch => updateTitle(e.target.value)}
+      />
+    </label>
 
-      <button type="button" onClick={() => addElement("1")}>
-        1
-      </button>
-      <button type="button" onClick={() => addElement("2")}>
-        2
-      </button>
-      <button type="button" onClick={() => addElement("3")}>
-        3
-      </button>
+    <button type="button" onClick={(): Dispatch => addElement('1')}>
+      1
+    </button>
+    <button type="button" onClick={(): Dispatch => addElement('2')}>
+      2
+    </button>
+    <button type="button" onClick={(): Dispatch => addElement('3')}>
+      3
+    </button>
 
-      <button type="button" onClick={deleteElement}>
-        Delete the last added
-      </button>
+    <button type="button" onClick={deleteElement}>
+      Delete the last added
+    </button>
 
-      {elements &&
-        elements.map(e => (
-          <p key={e.id}>
-            {e.title} {e.id}
-          </p>
-        ))}
-    </div>
-  );
-};
+    {elements &&
+      elements.map((e) => (
+        <p key={e.id}>
+          {e.title} {e.id}
+        </p>
+      ))}
+  </div>
+);
 
 export default connect(
   mapStateToProps,
-  mapDispatchToProps
+  mapDispatchToProps,
 )(Frontpage);
