@@ -1,4 +1,4 @@
-import React from "react";
+import React from 'react';
 
 const NotFound = (): JSX.Element => (
   <section className="page page--not-found">
